+++ src/app/cart/cart.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 
+import { SharedModule } from './../shared/shared.module';
 import { CartListComponent } from './components/cart-list/cart-list.component';
 import { CartItemComponent } from './components/cart-item/cart-item.component';
 
@@ -13,7 +13,7 @@ import { CartItemComponent } from './components/cart-item/cart-item.component';
     CartListComponent
   ],
   imports: [
-    CommonModule
+    SharedModule
   ],
 })
 export class CartModule { }