+++ src/app/core/services/console-print.service.ts
@@ -1,0 +1,8 @@
+import { Injectable } from '@angular/core';
+
+@Injectable()
+export class ConsolePrintService {
+  print(message: string): void {
+    console.log(message);
+  }
+}