+++ src/app/shared/shared.module.ts
@@ -1,11 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { HighlightDirective } from './directives/highlight.directive';
 
+import { HighlightDirective } from './directives/highlight.directive';
+import { ClickHandlerDirective } from './directives/click-handler.directive';
+import { OrderByPipe } from './pipes/order-by.pipe';
 
 @NgModule({
-  declarations: [HighlightDirective],
-  exports: [HighlightDirective],
+  declarations: [HighlightDirective, ClickHandlerDirective, OrderByPipe],
+  exports: [HighlightDirective, ClickHandlerDirective, OrderByPipe, CommonModule],
   imports: [
     CommonModule
   ]